pyproject.toml
[tool.poetry]
name = "beyond-crud"
version = "0.1.0"
description = ""
authors = ["riddler_rid <unkownrid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
sqlmodel = "^0.0.22"
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.0"
alembic = "^1.14.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
uvicorn = "^0.34.0"
pytest = "^8.3.4"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
